PRIMARY:
Integrate unit-testing into UnitySDK

Other changes:
Bugfixes (multiple)
Removed the AdminSDK temporarily since it's using features not supported by the new json library.
Adding some functionality to PlayFabHTTP to determine if there are pending messages.
Re-adding the DeveloperSecretKey to PlayFabSettings so the server api works again.  (This happened because we didn't have automated testing yet)
Pulling some UUnit improvements from CSharpSDK -> UnitySDK
The serverSDK folder has been converted to an example project (slightly renamed and adjusted the folder structure)
A new combinedSDK that is the primary unit-test project, and an example of everything merged.

Updating the unitypackage (which needs to be part of the automatic process eventually)
Updating the documentation to reflect the new 3-example project structure.

Stuff that wasn't properly added to the SDKGenerator the first time around:
Hiding the authKey (again).
Adding a call to test if the client is currently logged in (again).
Namespacing the unit tests (again).
Moving the UUnitTestRunner into an Editor folder (again).