Updating the DemoScene, to display slightly more information, and successfully login with the standard (non-device) environment.
Summary:
Updating the DemoScene, to display slightly more information, and successfully login with the standard (non-device) environment.
Adding the ExampleApiLog in DemoScene.
Adding a client-only version of the api-unit tests.
Updating the ios plugin to handle the api-callbacks that are going into the core sdk via SdkGenerator.

Test Plan: Client-only version of the usual api-tests.  Manual testing.

Reviewers: Marco.Williams, #dev_tools_team

Reviewed By: Marco.Williams, #dev_tools_team

Differential Revision: https://phab.playfabdev.com/D476