Fixed null reference error for account registration
Seems like there's a typo where the user settings were supposed to be loaded from the registration result. It's instead loaded from the login result, which has already been proven to be null due to the previous if-statement.
If `res` is null due to the result being a `RegisterPlayFabUserResult` instead of a `LoginResult` then the second time `AuthKey` is assigned, it will cause another null reference error. The assignment is redundant since one of the previous if-statements would have already assigned it.