Adding Quick Start menu
The plan here is to have a single button, you click it, given options on what you want this applied to, and accept the telemetry write we make.

NEXT STEP: this uses the static interface, we need to be able to make this with a new client api object instead.